Fix ingesting current block
Currently, the current block gets ignored if an event of interest
gets emitted in real time. This is because the ingestion engine
strictly requires the a filter's block range to be 1 or more.
This change ensures that the ingestion engine can query the current
block more than once depending on if the current block has already
been ingested or not.