Allow primary key columns in contract states
We have been building the migrations with the assumption that users
will not require a primary key in their contract states. As a result, the
build crashes when users specify a primary in their contract state migrations.
This commit attempts to fix this issue by retaining  the state version primary key
and removing the "PRIMARY KEY" keyword specified in the user migration when
building state versions migrations.