Introduce SharedState In EventHandlers (Breaking Change)
A shared state is exposed in event handlers to allow performing
side effects using the app's state or resources. A typical example
would be to store the DBPool for a web app in the shared state to
allow doing some DB related logic owned by your application.

When there is no shared state required for indexing, the user can simply
pass None in the `Config::new` build fn and `NoSharedState` in the
associated type for the event handlers.