Streamline events and contract addresses streams
This change ensures the queries used in streaming
through events and contract_addresses table
is composed respecting the originating chain and
contract (as it applies).

Since we still use diesel query builder, we had to
rewrite diesel streamer macros sloppily. This should
be alot cleaner after migrating to tokio-postgres
fully.