Allow optimizing chaindexing tasks
This change introduces KeepNodeActiveRequest, which is required
when enabling optimizations. In optimization mode, End users are
required to refresh the KeepNodeActiveRequest to allow Chaindexing
intelligently go inactive or active.

As a result, this will save users a bulk load of money by drastically reducing
RPC calls when there is no activity in their DApps.