Handle events in ContractAddresses Stream
This change removes handler subscriptions in favor of contract
addresses to cater for dynamic contracts use-cases.
Previously, handling was done chain-by-chain which created the
possibility of skipping newly created contract addresses.
Using the contract addresses as the new event subscription model,
it fixes that edge case and as a result, ergonomically allows handling
and ingesting events.